### 2024-04-03 #637 @realbp
cancerprof: API Client for extracting data from State Cancer Profiles
Submitting Author Name: Brian Park
Submitting Author Github Handle: @realbp
Repository:  https://github.com/getwilds/cancerprof
Version submitted: 0.1.0
Submission type: Standard

Version accepted: TBD

---

```
Package: cancerprof
Title: API Client for State Cancer Profiles
Version: 0.1.0
    person("Brian", "Park", , "joon.brianpark@gmail.com", role = c("aut", "cre"),
Description: An interface for retrieving data from the NIH NCI State Cancer Profiles API <https://statecancerprofiles.cancer.gov/>. State Cancer Profiles provides information about data topics including demographics, screening and risk factors, cancer incidence, and mortality for US states, counties, and health service areas.
    cdlTools,
    cli,
    httr2,
    utils

```

cancerprof allows users to retrieve data from [State Cancer Profiles](https://statecancerprofiles.cancer.gov) for programmable analysis. cancerprof makes accessing the undocumented API from State Cancer Profiles intuitive and easy.

The target audience for cancerprof is anyone who wants to access data from state cancer profiles to conduct programmable analysis without having to navigate the complex nature of its GUI. Specifically, cancer researchers could use cancerprof to conduct reproducable analysis of cancer crossed references with a variety of topics found within the data from state cancer profiles.

Currently there are no other softwares or packages that extracts the publicly available data from State Cancer Profiles.

Cancerprof does not breach any data privacy laws and complies with the ethics policies of ropensci.

https://github.com/ropensci/software-review/issues/635

Cancerprof passes all pkgcheck items
