+++ man/demo_crowding.Rd
@@ -44,7 +44,7 @@ demo_crowding(
   area = "WA",
   areatype = "county",
   crowding = "household with >1 person per room",
-  race = "All Races (includes Hispanic)", 
+  race = "All Races (includes Hispanic)",
   include_metadata = TRUE
 )
 
