### 2024-04-09 f8b9fe8
created get-metadata function
