+++ man/print.cancerprof_tbl.Rd
@@ -1,15 +1,20 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/process_metadata.R
+% Please edit documentation in R/process-metadata.R, R/process_metadata.R
 \name{print.cancerprof_tbl}
 \alias{print.cancerprof_tbl}
 \title{Custom print function}
 \usage{
+\method{print}{cancerprof_tbl}(x, ...)
+
 \method{print}{cancerprof_tbl}(x, ...)
 }
 \arguments{
 \item{x}{}
 }
 \description{
+This custom print function processes the
+metadata output for a response object
+
 This custom print function processes the
 metadata output for a response object
 }
