+++ NAMESPACE
@@ -39,5 +39,6 @@ importFrom(rlang,sym)
 importFrom(stats,setNames)
 importFrom(stringr,str_pad)
 importFrom(stringr,str_trim)
+importFrom(tibble,as_tibble)
 importFrom(utils,data)
 importFrom(utils,read.csv)
