### 2024-04-09 ece6f11
completed custom print for metadata
