### 2024-04-11 66b39c1
edited get metadata functions
