+++ man/extract_values.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get-metadata.R
+\name{extract_values}
+\alias{extract_values}
+\title{Extract Values}
+\usage{
+extract_values(key, resp_metadata)
+}
+\arguments{
+\item{key}{A String value}
+
+\item{resp_metadaa}{A list of strings of metadata}
+}
+\value{
+A string value without the key
+}
+\description{
+This function finds a string value that contains a key and removes the
+key from the value
+}
+\examples{
+\dontrun{
+extract_values("Sorted by", resp_metadata)
+}
+}
