+++ R/demo-population.R
@@ -114,7 +114,7 @@ demo_population <- function(area, areatype, population, race = NULL, sex = NULL)
 
   resp <- process_resp(resp, "demographics")
 
-  resp %>%
+  resp$data %>%
     setNames(c(
       get_area(areatype),
       "Percent",
