+++ R/demo-crowding.R
@@ -69,6 +69,8 @@ demo_crowding <- function(area, areatype, crowding, race) {
     ) %>%
     req_perform()
 
+  resp_url <- resp$url
+  
   resp <- process_resp(resp, "demographics")
   
   resp$data <- resp$data %>%
@@ -80,5 +82,5 @@ demo_crowding <- function(area, areatype, crowding, race) {
     )) %>%
     mutate(across(c("Percent", "Households"), \(x) as.numeric(x)))
 
-  process_metadata(resp)
+  process_metadata(resp, resp_url)
 }
