### 2024-04-16 #637 @ldecicco-USGS
---

There could be more information added to the README, although the bare minimum to meet our criteria is there.

In the examples I tried, my first thought was it might be nice to convert some of the text. For instance:

```r
x <- demo_income(
  area = "usa",
  areatype = "state",
  income = "median family income",
  race = "all races (includes hispanic)"
)
head(x$Rank)
[1] "52 of 52" "51 of 52" "50 of 52" "49 of 52" "48 of 52" "47 of 52"
```
Seems like c(52, 51, 50, etc) would be a more useful output to an R user. You'd probably want/need another column or _something_ to give the user the " of 52". Not mandatory, could be handy though (maybe a simple function to offer users outside of the function? or a simple example within the examples for how to extract the rank number).

---
