+++ pkgdown.yml
@@ -1,12 +1,12 @@
 pandoc: 3.1.11
-pkgdown: 2.0.7
+pkgdown: 2.0.8
 pkgdown_sha: ~
 articles:
   demographics-vignette: demographics-vignette.html
   incidence-vignette: incidence-vignette.html
   mortality-vignette: mortality-vignette.html
   risks-vignette: risks-vignette.html
-last_built: 2024-04-03T22:08Z
+last_built: 2024-04-17T18:18Z
 urls:
   reference: http://getwilds.org/cancerprof/reference
   article: http://getwilds.org/cancerprof/articles
