+++ NEWS.md
@@ -1,0 +1,3 @@
+# cancerprof 0.1.0
+
+* Initial CRAN submission.
