### 2024-04-17 #637 @realbp
Thank you for the feedback! I will make those changes in the upcoming version of cancerprof. I have added the ropensci badge and created a NEWS.md file.

What are the next steps in the review process?
