+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   incidence-vignette: incidence-vignette.html
   mortality-vignette: mortality-vignette.html
   risks-vignette: risks-vignette.html
-last_built: 2024-04-17T18:18Z
+last_built: 2024-04-17T18:50Z
 urls:
   reference: http://getwilds.org/cancerprof/reference
   article: http://getwilds.org/cancerprof/articles
