+++ R/get-metadata.R
@@ -34,36 +34,113 @@ print.cancerprof_metadata <- function(x, ...) {
   cli_div(theme = list(
     span.cancerprof_class = list(color = "darkgray")))
   
-  cli_text("{.cancerprof_class # Data Report:}")
-  for (i in seq_along(x$data_report)) {
-    cli_text(x$data_report[i], "\n")
+  data_topic <- attributes(x)$data_topic
+  
+  # do some conditionals to filter data topic
+  if (data_topic == "demographics") {
+    cli_text("{.cancerprof_class # Data Report:}")
+    for (i in seq_along(x$data_report)) {
+      cli_text(x$data_report[i], "\n")
+    }
+    cli_text("\n")
+    
+    cli_text("{.cancerprof_class # Sorted By:}")
+    cli_text(x$sortedby, "\n")
+    cli_text("\n")
+    
+    cli_text("{.cancerprof_class # Created By:}")
+    cli_text(x$createdby, "\n")
+    cli_text("\n")
+    
+    cli_text("{.cancerprof_class # Data Sources:}")
+    cli_text(x$data_sources, "\n")
+    cli_text("\n")
+    
+    cli_text("{.cancerprof_class # Data Dictionary:}")
+    cli_text(x$data_dictionary, "\n")
+    cli_text("\n")
+    
+    cli_text("{.cancerprof_class # Data Limitations:}")
+    cli_text(x$data_limitations, "\n")
+    cli_text("\n")
+    
+    if (!is.null(x$additional_notes) && length(x$additional_notes) > 0) {
+      cli_text("{.cancerprof_class # Additional Notes:}", "\n")
+      cli_text(x$additional_notes, "\n")
+    }
+  } else if (data_topic == "risks") {
+    
+  } else if (data_topic == "incidence" || data_topic == "mortality") {
+    cli_text("{.cancerprof_class # Data Report:}")
+    
+    for (i in seq_along(x$data_report)) {
+      cli_text(x$data_report[i], "\n")
+    }
+    cli_text("\n")
+    
+    cli_text("{.cancerprof_class # Sorted By:}")
+    cli_text(x$sortedby, "\n")
+    cli_text("\n")
+    
+    cli_text("{.cancerprof_class # Created By:}")
+    cli_text(x$createdby, "\n")
+    cli_text("\n")
+    
+    if (!is.null(x$trend) && length(x$x$trend) > 0) {
+      cli_text("{.cancerprof_class # Trend:}", "\n")
+      cli_text(x$x$trend, "\n")
+      cli_text("\n")
+    }
+    
+    if (!is.null(x$trend_note) && length(x$trend_note) > 0) {
+      cli_text("{.cancerprof_class # Trend Note:}", "\n")
+      cli_text(x$trend_note, "\n")
+      cli_text("\n")
+    }
+    
+    if (!is.null(x$rate_note) && length(x$rate_note) > 0) {
+      cli_text("{.cancerprof_class # Rate Note:}", "\n")
+      cli_text(x$rate_note, "\n")
+      cli_text("\n")
+    }
+    
+    if (!is.null(x$stage_note) && length(x$stage_note) > 0) {
+      cli_text("{.cancerprof_class # Stage Note:}", "\n")
+      cli_text(x$stage_note, "\n")
+      cli_text("\n")
+    }
+    
+    if (!is.null(x$rank_note) && length(x$rank_note) > 0) {
+      cli_text("{.cancerprof_class # Rank Note:}", "\n")
+      cli_text(x$rank_note, "\n")
+      cli_text("\n")
+    }
+    
+    if (!is.null(x$data_not_available) && length(x$data_not_available) > 0) {
[…]
