### 2024-04-19 86ebc9e
moved custom print functions and extract_values function to utils.R
