+++ R/get-metadata.R
@@ -1,161 +1,3 @@
-#' Extract Values
-#'
-#' This function finds a string value that contains a key and removes the
-#' key from the value
-#'
-#' @param key A String value
-#' @param resp_metadaa A list of strings of metadata
-#' 
-#' @returns A string value without the key
-#' 
-#' @examples
-#' \dontrun{
-#' extract_values("Sorted by", resp_metadata)
-#' }
-extract_values <- function(key, resp_metadata) {
-  values <- resp_metadata[grep(key, resp_metadata)]
-  values <- gsub(paste0("^\\s*", key, ":?\\s*"), "", values)
-  return(values)
-}
-
-#' Custom print function
-#'
-#' This custom print function processes the
-#' metadata output for easier readability
-#'
-#' @param x
-#'
-#' @export
-print.cancerprof_metadata <- function(x, ...) {
-  
-  cli_h1("Metadata")
-  cli_text("\n")
-  
-  cli_div(theme = list(
-    span.cancerprof_class = list(color = "darkgray")))
-  
-  data_topic <- attributes(x)$data_topic
-  
-  # do some conditionals to filter data topic
-  if (data_topic == "demographics" || data_topic == "risks") {
-    cli_text("{.cancerprof_class # Data Report:}")
-    for (i in seq_along(x$data_report)) {
-      cli_text(x$data_report[i], "\n")
-    }
-    cli_text("\n")
-    
-    cli_text("{.cancerprof_class # Sorted By:}")
-    cli_text(x$sortedby, "\n")
-    cli_text("\n")
-    
-    cli_text("{.cancerprof_class # Created By:}")
-    cli_text(x$createdby, "\n")
-    cli_text("\n")
-    
-    if (!is.null(x$data_sources) && length(x$data_sources) > 0) {
-      cli_text("{.cancerprof_class # Data Sources:}", "\n")
-      cli_text(x$data_sources, "\n")
-      cli_text("\n")
-    }
-    
-    if (!is.null(x$data_dictionary) && length(x$data_dictionary) > 0) {
-      cli_text("{.cancerprof_class # Data Dictionary:}", "\n")
-      cli_text(x$data_dictionary, "\n")
-      cli_text("\n")
-    }
-    
-    if (!is.null(x$data_limitations) && length(x$data_limitations) > 0) {
-      cli_text("{.cancerprof_class # Data Limitations:}", "\n")
-      cli_text(x$data_limitations, "\n")
-      cli_text("\n")
-    }
-    
-    if (!is.null(x$name_change) && length(x$name_change) > 0) {
-      cli_text("{.cancerprof_class # Name Change:}", "\n")
-      cli_text(x$name_change, "\n")
-      cli_text("\n")
-    }
-    
-    if (!is.null(x$additional_notes) && length(x$additional_notes) > 0) {
-      cli_text("{.cancerprof_class # Additional Notes:}", "\n")
-      cli_text(x$additional_notes, "\n")
-    }
-    
-  } else if (data_topic == "incidence" || data_topic == "mortality") {
-    cli_text("{.cancerprof_class # Data Report:}")
-    
-    for (i in seq_along(x$data_report)) {
-      cli_text(x$data_report[i], "\n")
-    }
-    cli_text("\n")
-    
-    cli_text("{.cancerprof_class # Sorted By:}")
-    cli_text(x$sortedby, "\n")
-    cli_text("\n")
-    
-    cli_text("{.cancerprof_class # Created By:}")
-    cli_text(x$createdby, "\n")
-    cli_text("\n")
-    
-    if (!is.null(x$trend) && length(x$x$trend) > 0) {
[…]
