+++ man/print.cancerprof_metadata.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get-metadata.R
+% Please edit documentation in R/utils.R
 \name{print.cancerprof_metadata}
 \alias{print.cancerprof_metadata}
 \title{Custom print function}
@@ -11,5 +11,5 @@
 }
 \description{
 This custom print function processes the
-metadata output for easier readability
+metadata output for each data topic for better readability
 }
