### 2024-04-22 c5766d1
created pretty print functionality
