### 2024-04-23 1b0f314
created and completed trend-fips-scp for trend area data pulling
