### 2024-04-24 ea4a75f
able to pull trend data
