### 2024-04-24 74ac0ba
pull-trend-data works properly with metadata attribute
