+++ R/pull-trend-data.R
@@ -1,40 +1,82 @@
-
+#' Access to Trend Data
+#'
+#' This function returns a data frame about trend data on incidence and
+#' mortality from State Cancer Profiles
+#'
+#' @param area A state/territory abbreviation or USA or "seer 9 registeries"
+#' @param cancer One of the following values:
+#' - `"all cancer sites"`
+#' - `"bladder"`
+#' - `"brain & ons"`
+#' - `"breast (female)"`
+#' - `"breast (female in situ)"`
+#' - `"cervix"`
+#' - `"childhood (ages <15, all sites)"`
+#' - `"childhood (ages <20, all sites)"`
+#' - `"colon & rectum"`
+#' - `"esophagus"`
+#' - `"kidney & renal pelvis"`
+#' - `"leukemia"`
+#' - `"liver & bile duct"`
+#' - `"lung & bronchus"`
+#' - `"melanoma of the skin"`
+#' - `"non-hodgkin lymphoma"`
+#' - `"oral cavity & pharynx"`
+#' - `"ovary"`
+#' - `"pancreas"`
+#' - `"prostate"`
+#' - `"stomach"`
+#' - `"thyroid"`
+#' - `"uterus (corpus & uterus, nos)"`.
+#' @param race
+#' One of the following values:
+#' - `"all Races (includes Hispanic)"`
+#' - `"white non-hispanic"`
+#' - `"black (non-hispanic)"`
+#' - `"american indian / alaska native (non-hispanic)"`
+#' - `"asian / pacific islander (non-hispanic)"`
+#' - `"hispanic (Any Race)"`.
+#' @template param-sex
+#' @param age One of the following values:
+#' - `"all ages"`
+#' - `"ages <50"`
+#' - `"ages 50+"`
+#' - `"ages <65"`
+#' - `"ages 65+"`
+#' @param datatype One of the following values:
+#' - `"incidence"`
+#' - `"mortality"`
+#'
+#' @importFrom httr2 req_url_query req_perform
+#' @importFrom stats setNames
+#' @importFrom dplyr mutate across
+#' @importFrom tibble as_tibble
+#'
+#' @returns A data frame with the following columns: Year, Observed, Estimated
+#'
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' pull_trend_data(
+#'   area = "wa",
+#'   cancer = "lung & bronchus",
+#'   data = "incidence",
+#'   race = "All Races (includes Hispanic)",
+#'   sex = "both sexes",
+#'   age = "all ages"
+#' )
+#' }
 pull_trend_data <- function(area, cancer, race, sex, age, datatype) {
-
-  
-  # url <- "https://statecancerprofiles.cancer.gov/historicaltrend/data.php/historicaltrend.csv"
-  url <- "https://statecancerprofiles.cancer.gov/historicaltrend/data.php/historicaltrend.csv?0&9900&999&7599&001&001&00&0&0&0&1&0&1&1&6"
-
-  req <- request(url)
   
   req <- create_request("trend")
-  
-  #?0&9900&999&7599&001&001&00&0&0&0&1&0&1&1&6
-  #req_url_path_append()
-  
-  # seems to be always "0" if no new lines are added
-  # number of additional trend lines
-  # "99" + FIPS or "5099 for SEER 9, UTAH = 2649 NEW YORK = 4636, NEW MEXICO = 2335, NEW JERSEY = 4434, MASSACHUSETTS = 4725, LOUISIANA = 4322, KENTUCKY = 4221, IOWA = 2219, IDAHO = 4516, HAWAII = 2115, CONNECTICUT = 0209, CALIFORNIA = 9706 
-  # 999
-  # 7599
-  # age
-  # cancer
-  # race
-  # sex
-  # 
-  # 
[…]
