+++ man/print.cancerprof_tbl.Rd
@@ -1,20 +1,15 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/process-metadata.R, R/utils.R
+% Please edit documentation in R/utils.R
 \name{print.cancerprof_tbl}
 \alias{print.cancerprof_tbl}
 \title{Custom print function}
 \usage{
-\method{print}{cancerprof_tbl}(x, ...)
-
 \method{print}{cancerprof_tbl}(x, ...)
 }
 \arguments{
 \item{x}{}
 }
 \description{
-This custom print function processes the
-metadata output for a response object
-
 This custom print function edits the comment on the
 metadata tibble output for a response object
 }
