### 2024-04-25 0979343
get-metadata, link to scp works with trend
