+++ man/process_metadata.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/process_metadata.R
+% Please edit documentation in R/process-metadata.R
 \name{process_metadata}
 \alias{process_metadata}
 \title{Process Metadata}
