### 2024-04-26 2daff5b
moved get-raw-metadata to utils, edited documentation
