+++ man/print.cancerprof_metadata.Rd
@@ -1,15 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{print.cancerprof_metadata}
-\alias{print.cancerprof_metadata}
-\title{Custom print function}
-\usage{
-\method{print}{cancerprof_metadata}(x, pretty_print = TRUE, ...)
-}
-\arguments{
-\item{x}{}
-}
-\description{
-This custom print function processes the
-metadata output for each data topic for better readability
-}
