+++ NAMESPACE
@@ -1,7 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(print,cancerprof_metadata)
-S3method(print,cancerprof_tbl)
 export("%>%")
 export(demo_crowding)
 export(demo_education)
