### 2024-04-29 59b1e3c
edited pull_trend_data to historical-trends, minor comments
