+++ man/historical_trends.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/pull-trend-data.R
-\name{pull_trend_data}
-\alias{pull_trend_data}
+\name{historical_trends}
+\alias{historical_trends}
 \title{Access to Trend Data}
 \usage{
-pull_trend_data(area, cancer, race, sex, age, datatype)
+historical_trends(area, cancer, race, sex, age, datatype)
 }
 \arguments{
 \item{area}{A state/territory abbreviation or USA or "seer 9 registeries"}
@@ -77,7 +77,7 @@ mortality from State Cancer Profiles
 }
 \examples{
 \dontrun{
-pull_trend_data(
+historical_trends(
   area = "wa",
   cancer = "lung & bronchus",
   race = "All Races (includes Hispanic)",
@@ -86,7 +86,7 @@ pull_trend_data(
   datatype = "incidence"
 )
 # no data available
-pull_trend_data(
+historical_trends(
   area = "wa",
   cancer = "lung & bronchus",
   race = "black (non-hispanic)",
