+++ NAMESPACE
@@ -12,9 +12,9 @@ export(demo_population)
 export(demo_poverty)
 export(demo_svi)
 export(demo_workforce)
+export(historical_trends)
 export(incidence_cancer)
 export(mortality_cancer)
-export(pull_trend_data)
 export(risk_alcohol)
 export(risk_colorectal_screening)
 export(risk_diet_exercise)
@@ -30,6 +30,7 @@ importFrom(dplyr,mutate)
 importFrom(dplyr,mutate_all)
 importFrom(dplyr,na_if)
 importFrom(httr2,req_perform)
+importFrom(httr2,req_url_path_append)
 importFrom(httr2,req_url_query)
 importFrom(httr2,request)
 importFrom(httr2,resp_body_string)
