+++ man-roxygen/param-sex.R
@@ -1,4 +1,4 @@
 #' @param sex One of the following values:
 #' - `"both sexes"`
-#' - `"male"`
-#' - `"female"`.
+#' - `"males"`
+#' - `"females"`.
