### 2024-04-29 bcaba51
first attempt to fix printing issue
