+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,cancerprof_tbl)
 export("%>%")
 export(demo_crowding)
 export(demo_education)
@@ -23,6 +24,9 @@ export(risk_vaccines)
 export(risk_women_health)
 importFrom(cdlTools,fips)
 importFrom(cli,cli_abort)
+importFrom(cli,cli_par)
+importFrom(cli,cli_text)
+importFrom(cli,make_ansi_style)
 importFrom(dplyr,across)
 importFrom(dplyr,all_of)
 importFrom(dplyr,filter)
