+++ man/risk_diet_exercise.Rd
@@ -31,8 +31,8 @@ risk_diet_exercise(diet_exercise, race, sex)
 \item{sex}{One of the following values:
 \itemize{
 \item \code{"both sexes"}
-\item \code{"male"}
-\item \code{"female"}.
+\item \code{"males"}
+\item \code{"females"}.
 }}
 }
 \value{
