+++ man/historical_trends.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pull-trend-data.R
+% Please edit documentation in R/historical-trends.R
 \name{historical_trends}
 \alias{historical_trends}
-\title{Access to Trend Data}
+\title{Access to Historical Trends Data}
 \usage{
 historical_trends(area, cancer, race, sex, age, datatype)
 }
 \arguments{
-\item{area}{A state/territory abbreviation or USA or "seer 9 registeries"}
+\item{area}{A state/territory abbreviation or USA or "seer 9 registries"}
 
 \item{cancer}{One of the following values:
 \itemize{
@@ -49,8 +49,8 @@ historical_trends(area, cancer, race, sex, age, datatype)
 \item{sex}{One of the following values:
 \itemize{
 \item \code{"both sexes"}
-\item \code{"male"}
-\item \code{"female"}.
+\item \code{"males"}
+\item \code{"females"}.
 }}
 
 \item{age}{One of the following values:
