+++ man/demo_population.Rd
@@ -49,8 +49,8 @@ demo_population(area, areatype, population, race = NULL, sex = NULL)
 \item{sex}{One of the following values:
 \itemize{
 \item \code{"both sexes"}
-\item \code{"male"}
-\item \code{"female"}.
+\item \code{"males"}
+\item \code{"females"}.
 }}
 }
 \value{
