+++ man/risk_vaccines.Rd
@@ -16,8 +16,8 @@ risk_vaccines(vaccine, sex)
 \item{sex}{One of the following values:
 \itemize{
 \item \code{"both sexes"}
-\item \code{"male"}
-\item \code{"female"}.
+\item \code{"males"}
+\item \code{"females"}.
 }}
 }
 \value{
