+++ NAMESPACE
@@ -23,6 +23,9 @@ export(risk_vaccines)
 export(risk_women_health)
 importFrom(cdlTools,fips)
 importFrom(cli,cli_abort)
+importFrom(cli,cli_par)
+importFrom(cli,cli_text)
+importFrom(cli,make_ansi_style)
 importFrom(dplyr,across)
 importFrom(dplyr,all_of)
 importFrom(dplyr,filter)
