+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,cancerprof_metadata)
 S3method(print,cancerprof_tbl)
 export("%>%")
 export(demo_crowding)
@@ -24,6 +25,8 @@ export(risk_vaccines)
 export(risk_women_health)
 importFrom(cdlTools,fips)
 importFrom(cli,cli_abort)
+importFrom(cli,cli_div)
+importFrom(cli,cli_h1)
 importFrom(cli,cli_par)
 importFrom(cli,cli_text)
 importFrom(cli,make_ansi_style)
