### 2024-04-30 #637 @ldecicco-USGS
Do you get the same error if you install terra and raster independently?

```r
install.packages(c("terra", "raster"))
```
