### 2024-05-02 #637 @jromanowska
Today I've tried on another computer (also Linux) and the `pak` command worked - it actually showed me which system libraries were missing :thinking:
After installing those, I could re-run the `pak` without problems. I will write some comments about this installation process and issues in my review, so that other users may be aware.
