### 2024-05-02 646a242
fixed recent trend issue in get-metadata
