+++ NAMESPACE
@@ -26,9 +26,11 @@ export(risk_women_health)
 importFrom(cdlTools,fips)
 importFrom(cli,cli_abort)
 importFrom(cli,cli_div)
+importFrom(cli,cli_end)
 importFrom(cli,cli_h1)
 importFrom(cli,cli_par)
 importFrom(cli,cli_text)
+importFrom(cli,cli_vec)
 importFrom(cli,make_ansi_style)
 importFrom(dplyr,across)
 importFrom(dplyr,all_of)
