### 2024-05-02 7febc44
created test files for metadata and trends
