+++ tests/testthat/test-get-metadata.R
@@ -1,0 +1,4 @@
+#' Test Get Metadata
+#'
+#' This testthat file test the get-metadata function
+
