+++ tests/testthat/test-historical-trends.R
@@ -1,0 +1,4 @@
+#' Test Historical Trends
+#'
+#' This testthat file test the historical-trends function
+
