+++ tests/testthat/test-trend-fips-scp.R
@@ -1,0 +1,4 @@
+#' Test Trend FIPS SCP
+#'
+#' This testthat file test the trend-fips-scp function
+
