+++ tests/testthat/test-trend-fips-scp.R
@@ -1,4 +1,3 @@
 #' Test Trend FIPS SCP
 #'
 #' This testthat file test the trend-fips-scp function
-
