+++ tests/testthat/test-handle_trend_datatype.R
@@ -1,3 +1,3 @@
-test_that("multiplication works", {
-  expect_equal(2 * 2, 4)
-})
+#' Test Handle Trend Datatype
+#'
+#' This testthat file test the handle-trend-datatype function
