### 2024-05-24 #637 @jromanowska
Thanks for the patience! Great work with {cancerprof}! :raised_hands:

I'm done with reviewing, please check my comments below.

No working relationship.

  - not really - the authors could expand the description by adding 1-2 sentences;
  - there is no order in vignettes/articles, so it's difficult to navigate the
  help page

  - the required system libraries could be written down (e.g., on Linux, I had
  to install some separately before installing the package);
  - while `pak` is a great package, some people might not use it and try installing
  `cancerprof` via `devtools` which might fail not giving useful information about
  source of error - the authors could add why they recommend installing via `pak`

  - I could not launch vignettes locally - it says `no vignettes found` when
  installed
  - I tried reading the vignettes on the webpage - I have some comments to those,
  please check [below](#is-the-documentation-clear-and-sufficient)

  - almost all examples are `not run` - it would be more useful if each example
  showed actually that the function works; this would also make the documentation
  better

  - README could be more detailed, as mentioned above
  - webpage is nice (although the vignettes should be changed, see the comments below)
  - when checking with `devtools::check()` I get a warning:
  `Warning: program compiled against libxml 210 using older 209` - I'm not sure
  where it comes from
  - tests were nicely written, using `testthat`, but maybe the `test-dput*R`
  files that are in the `R` directory should be in the `tests`?

---

## Review Comments

Overall, the package is really useful for batch retrieval of data.
The functions are well documented and designed, tests are well written.
I have some comments that I hope will improve the usability of both, the
package and documentation.

### Are there improvements that could be made to the code style?

1. Avoid long lines, e.g., lines 59-63 in `demo-education.R` could be split
after each logical operator. This will also improve readability of the if
clauses.
    - run `goodpractice::gp()` to get more tips

2. Running the `styler` package showed that only 4 files could be improved -
great work! You don't need to, but you can try that on these files:
    - R/risk-womens-health.R
    - R/risk-smoking.R
    - R/demo-poverty.R
    - R/demo-population.R

### Is there code duplication in the package that should be reduced?

1. When dealing with specific text, like the various levels of smoking, try to
gather these in one place.
Currently, these various levels of smoking are
defined in both `risk-smoking.R` and `handle-smoking.R`. This is difficult to
maintain. One way to deal with that is to create an environment that is loaded
when the package is loaded. Then, you can just grab the entire objects from
this environment, without the need for the user to be bothered by these objects.
[Check here](https://r-pkgs.org/code.html#sec-code-onLoad-onAttach) for some
more explanation.

### Are there user interface improvements that could be made?

1. Use the default values of a function to ease function calling
    - if there is only one possible value for a certain argument, implement this
  as a default value (I understand that it might change in the future but
  it could be default without any problem), e.g., `demo_language` _requires_
  the user to write `language = "language isolation"` even if it's the only
  possible value
    - the functions that change their required arguments when one of the
  arguments is chosen are both, difficult to use and difficult to maintain -
  perhaps you can just give a warning that some arguments are omitted?
    - e.g., in `demo_education`, instead of aborting when the user does not provide some
  arguments, one can give a warning and assume _default_ values
    - especially in [`Risk smoking`](https://getwilds.org/cancerprof/articles/risks-vignette.html#risk-smoking)
  there is so many various combinations that it might be difficult to use that
  function when programmatically fetching data - perhaps you could split this
  into smaller, more focused functions?
    - similarly, for some arguments combinations, another argument is required
  to be only of a certain value (e.g., `incidence_cancer` with `cancer = ovary`
  is possible only with `sex = "females"`), thus in such cases, if the user
  provides a wrong argument the function might just return the correct table and
  issue a warning

1. The `area` and `areatype` arguments are not checked anywhere, so when user
puts a wrong value, they get an uninformative error:
```
Error in `process_resp()`:
! Invalid input, please check documentation for valid arguments.
Run `rlang::last_trace()` to see where the error occurred.
```

### Are there performance improvements that could be made?

No issues here.
The performace depends here on the speed of internet connection of the user.

### Is the documentation clear and sufficient?

1. I got error when trying to install the package locally, from the cloned GitHub repo:
`gdal-config not found` - I can't find any mention that any software like
`gdal` is reqiured by `cancerprof`
2. trying to install from the net, using `pak`, as the instructions say, I get the error:

```
pak::pak("getwilds/cancerprof")
! error in pak subprocess
Caused by error:
! Could not solve package dependencies:
* getwilds/cancerprof: ! pkgdepends resolution error for getwilds/cancerprof.
Caused by error:
! Bad GitHub credentials, make sure that your GitHub token is valid.
Caused by error in `stop(http_error(resp))`:
! Unauthorized (HTTP 401).
Type .Last.error to see the more details.
```
_(I got this resolved afterwards)_

#### Vignettes and function documentation

1. The vignettes are basically a list of examples for each of the functions.
These examples should be (and in most cases are) in the function documentation. Instead,
a vignette should guide a user through a specific use case scenario.
Could the authors create a set of scientific questions that the package helps to
answer? Especially batch retrieval of data is important.
1. I could not find explanation of "FIPS"
1. In "Demo mobility" it says: "The function defaults to `"all races"`,
`"both sexes"`, `"ages 1+"`" - what does it mean? That the data from the source
is available only for these categories?
Usually, when I see "defaults", I would assume that these are the values
assumed _if_ there is no specific value for an argument. Here, it seems like
there can't be any other value.

1. There is a mistake in `risk_women_health` documentation - wrong names of the
columns of returned data frame.
1. The authors could explain the different categories of arguments more, e.g.,
in `risk_women_health`, the `women_health` argument has three possible values
("pap smear in past 3 years, no hysterectomy, ages 21-65", "mammogram in past 2 years, ages 50-74", "mammogram in past 2 years, ages 40+") and the columns of the returned data frame are not informative
enough to understand what type of data one actually gets

1. Is there a typo in `risk_alcohol` documentation, under `alcohol` argument?
(and does this argument value need to be so long?)

#### Does the documentation use the principle of multiple points of entry i.e. takes into account the fact that any piece of documentation may be the first encounter the user has with the package and/or the tool/data it wraps?

1. The webpage is well built, so that one can get to the home page easily
1. However, there is no common help-page for the package (i.e.,
  `help(package = "cancerprof")` gives
  `URL '/help/library/cancerprof/html/00Index.html' not found`)

### Were functions and arguments named to work together to form a common, logical programming API that is easy to read, and autocomplete?

Yes - good choice of naming. Although the `demo_` functions sounded at first
like "demonstration" to me, but it's logical to abbreviate "demography" as "demo". :)
