### 2024-06-05 fb343b8
added error message for invalid area input
