### 2024-06-05 899fb12
created handle-areatype to check invalid inputs. added handle-areatype to crowding
