### 2024-06-12 #637 @ginberg
Well done with creating this package!
See below for my comments.

The README is very brief. Please add some explanation about what type of data can be obtained. Also can you mention the target audience?
I had no problems installing the package
It's great that there is a vignette about each data topic. What I miss is an overview of State Cancer Profiles and how these vignettes are connected. I think an overview vignette that contains a description about State Cancer Profiles with links to the other vignettes is useful. In this vignette you could also mention which type of user could use a certain vignette.
Each exported function has examples but they are mostly inside a dontrun, so they don't actually run. It would be useful if at least some of them can be run locally.

---

- Does the code comply with general principles in the Mozilla reviewing guide?
If there is a parameter in a function that can only have 1 value (e.g. 'crowding' in demo_crowding), you don't need to make it a parameter at all. Or set the value as default. Try to make it as simple as possible for the user to use the function.
- Is there code duplication in the package that should be reduced?
There is some code duplication, especially related to texts. For example the text in handle_alcohol. Better make it a variable so it will be easier to update.
- Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient?
the README and vignettes could be improved (see above)
- Other
The package relies heavily on a working internet connection, if it's not available the tests will fail.
```
ℹ Testing cancerprof
✔ | F W  S  OK | Context
✖ | 3        1 | demo-crowding
────────────────────────────────────────────────────────────────────
Error (test-demo-crowding.R:8:3): Output data type is correct
<httr2_failed/rlang_error/error/condition>
Error: Could not resolve host: statecancerprofiles.cancer.gov
    ▆
 1. ├─cancerprof::demo_crowding(...) at test-demo-crowding.R:8:2
 2. │ └─... %>% req_perform() at cancerprof/R/demo-crowding.R:57:2
 3. └─httr2::req_perform(.)
 4.   └─base::tryCatch(...)
 5.     └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 6.       └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 7.         └─value[[3L]](cond)
```
 It would be good to fail gracefully with an informative message if a resource is not available (and not give a error). See [here](https://forum.posit.co/t/internet-resources-should-fail-gracefully/49199) for some ideas. You probably need to fix this anyway when you want the package to be on CRAN.
