+++ R/demo-mobility.R
@@ -44,9 +44,8 @@
 #' )
 #' }
 demo_mobility <- function(area, areatype, mobility) {
-  req <- create_request("demographics")
-
-  resp <- req %>%
+  # Request
+  req <- create_request("demographics") %>% 
     req_url_query(
       stateFIPS = fips_scp(area),
       areatype = tolower(areatype),
@@ -55,13 +54,13 @@ demo_mobility <- function(area, areatype, mobility) {
       type = "manyareacensus",
       sortVariableName = "value",
       sortOrder = "default",
-      output = 1
-    ) %>%
-    req_perform()
-
+      output = 1)
+  
+  # Response
+  resp <- req_perform(req)
   resp <- process_resp(resp, "demographics")
-
-  resp %>%
+  
+  resp$data %>%
     setNames(c(
       get_area(areatype),
       "Percent",
