+++ R/demo-svi.R
@@ -39,9 +39,8 @@
 #' )
 #' }
 demo_svi <- function(area, svi) {
-  req <- create_request("demographics")
-
-  resp <- req %>%
+  # Request
+  req <- create_request("demographics") %>% 
     req_url_query(
       stateFIPS = fips_scp(area),
       areatype = "county",
@@ -51,9 +50,10 @@ demo_svi <- function(area, svi) {
       sortVariableName = "value",
       sortOrder = "default",
       output = 1
-    ) %>%
-    req_perform()
-
+    )
+  
+  # Response
+  resp <- req_perform(req)
   resp <- process_resp(resp, "demographics")
 
   resp$data %>%
