### 2024-07-08 860a267
First look at rest of the functions
