+++ R/risk-alcohol.R
@@ -39,9 +39,7 @@
 #' )
 #' }
 risk_alcohol <- function(alcohol, race, sex) {
-  req <- create_request("risk")
-
-  resp <- req %>%
+  req <- create_request("risk") %>% 
     req_url_query(
       topic = "alcohol",
       risk = handle_alcohol(alcohol),
@@ -51,12 +49,12 @@ risk_alcohol <- function(alcohol, race, sex) {
       sortVariableName = "percent",
       sortOrder = "default",
       output = 1
-    ) %>%
-    req_perform()
-
+    )
+  
+  resp <- req_perform(req)
   resp <- process_resp(resp, "risks")
-
-  resp %>%
+  
+  resp$data %>%
     setNames(c(
       "State",
       "FIPS",
