### 2024-07-09 9a38656
Add `process_metadata()` at the end
