+++ R/demo-insurance.R
@@ -121,9 +121,12 @@ demo_insurance <- function(area, areatype, insurance, sex, age, race = NULL) {
   
   # Response
   resp <- req_perform(req)
+  resp_url <- resp$url
   resp <- process_resp(resp, "demographics")
   
-  resp$data %>%
+  resp$data <- resp$data %>%
     setNames(c(get_area(areatype), "Percent", "People", "Rank")) %>%
     mutate(across(c("Percent", "People"), \(x) as.numeric(x)))
+  
+  process_metadata(resp, "demographics", resp_url)
 }
