+++ R/demo-svi.R
@@ -54,11 +54,12 @@ demo_svi <- function(area, svi) {
   
   # Response
   resp <- req_perform(req)
+  resp_url <- resp$url
   resp <- process_resp(resp, "demographics")
-
-  resp$data %>%
+  
+  resp$data <- resp$data %>%
     setNames(c("County", "FIPS", "Score")) %>%
     mutate(across(c("Score"), \(x) as.numeric(x)))
   
-  process_metadata(resp)
+  process_metadata(resp, "demographics", resp_url)
 }
