+++ R/demo-workforce.R
@@ -71,9 +71,10 @@ demo_workforce <- function(area, areatype, workforce, race, sex) {
   
   # Response
   resp <- req_perform(req)
+  resp_url <- resp$url
   resp <- process_resp(resp, "demographics")
-
-  resp$data %>%
+  
+  resp$data <- resp$data %>%
     setNames(c(
       get_area(areatype),
       "Percent",
@@ -81,4 +82,6 @@ demo_workforce <- function(area, areatype, workforce, race, sex) {
       "Rank"
     )) %>%
     mutate(across(c("Percent", "People_Unemployed"), \(x) as.numeric(x)))
+  
+  process_metadata(resp, "demographics", resp_url)
 }
