+++ R/demo-non-english-language.R
@@ -55,9 +55,10 @@ demo_language <- function(area, areatype, language) {
   
   # Response
   resp <- req_perform(resp)
+  resp_url <- resp$url
   resp <- process_resp(resp, "demographics")
   
-  resp$data %>%
+  resp$data <- resp$data %>%
     setNames(c(
       get_area(areatype),
       "Percent",
@@ -65,4 +66,6 @@ demo_language <- function(area, areatype, language) {
       "Rank"
     )) %>%
     mutate(across(c("Percent", "Households"), \(x) as.numeric(x)))
+  
+  process_metadata(resp, "demographics", resp_url)
 }

[…]
