### 2024-07-30 #637 @seankross
Hi @ldecicco-USGS, I supervised @realbp's work on this package as part of his internship at the Fred Hutch Data Science Lab, which has sadly ended. We loved working with Brian! Would it be possible for me to take reigns of this review? I plan to be the maintainer of this package once we eventually submit it to CRAN.
