+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: cancerprof
 Title: API Client for State Cancer Profiles
-Version: 0.1.0
+Version: 0.1.0.9001
 Authors@R: 
     person("Brian", "Park", , "joon.brianpark@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0009-0008-8274-3057"))
@@ -10,7 +10,7 @@ URL: https://github.com/getwilds/cancerprof, https://getwilds.org/cancerprof/
 BugReports: https://github.com/getwilds/cancerprof/issues
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.1
+RoxygenNote: 7.3.2
 Imports: 
     cdlTools,
     cli,
