+++ NAMESPACE
@@ -38,6 +38,8 @@ importFrom(dplyr,filter)
 importFrom(dplyr,mutate)
 importFrom(dplyr,mutate_all)
 importFrom(dplyr,na_if)
+importFrom(dplyr,rename)
+importFrom(dplyr,select)
 importFrom(httr2,req_perform)
 importFrom(httr2,req_url_path_append)
 importFrom(httr2,req_url_query)
@@ -50,5 +52,6 @@ importFrom(stats,setNames)
 importFrom(stringr,str_pad)
 importFrom(stringr,str_trim)
 importFrom(tibble,as_tibble)
+importFrom(tidyr,separate_wider_regex)
 importFrom(utils,data)
 importFrom(utils,read.csv)
