### 2025-02-02 #637 @maurolepore
Dear @realbp and @seankross this is to mark the start of my EiC rotation. I'm reviewing all open issues and leaving a short note to myself of what I see.

I see @seankross replaced @realbp in the role of the author. Both reviews are in and we're waiting for the author to respond but it's taking a while so the previous EiC put it on hold.

@seankross could you please let us know if you intend to resume or stay on hold? It's totally OK to stay on hold since we understand priorities can change. But please do touch base to avoid closing the submission for lack of response.

---

https://devguide.ropensci.org/softwarereview_policies.html#policiesreviewprocess

> * If the author hasn’t requested a holding label, but is simply not responding, we should close the issue within one month after the last contact intent.
