### 2025-05-04 #637 @maurolepore
Dear @realbp and @seankross I'm wrapping up my EiC rotation and I see no activity since I started. We should probabbly close. I emailed @seankross for confirmation.

In any case, I hope you're well and thanks a lot for sharing your work with rOpenSci. I look forward to working with you again whenever it's a good time.
