### 2026-05-03 #637 @maurolepore
Dear @seankross, this is to mark the start of my new EiC rotation.

A brief summary of what I see:
- Both reviews have been in for over a year.
- You confirmed in May 2025 that you intended to follow up.
- Since then there have been a few EiC check-ins, but no response on this thread.

I'd like to acknowledge your stated intent to continue, and at the same time the amount of time that has passed without an update. For context, our policy says:

> If the author hasn't requested a holding label, but is simply not responding, we should close the issue within one month after the last contact intent.
> — https://devguide.ropensci.org/softwarereview_policies.html#policiesreviewprocess

I'll ping you on Slack and email and look forward to hearing from you within a month.

Thanks!
